assign controllers for each resource
