+++ app/Http/Controllers/coursesController.php
@@ -13,147 +13,8 @@ use App\Models\Sections;
 use App\Models\Courses;
 use Symfony\Component\Stopwatch\Section;
 
-class resourcesController extends Controller
+class coursesController extends Controller
 {
-    //FUNCTIONS COURSES
-
-    public function listDegrees()
-    {
-        $search = request()->query('search');
-
-        if ($search) {
-            $degree = Degree::where('name', 'LIKE', "%{$search}%")
-                ->where('state', 1)
-                ->paginate(10);
-        } else {
-            $degree = Degree::where('state', 1)->paginate(10);
-        }
-
-        return view('resourcesJV.degrees.listDegrees', ['degree' => $degree]);
-    }
-
-    public function createDegrees(DegreeRequest $request)
-    {
-        $degree = Degree::create($request->validated());
-        return redirect('/degrees');
-    }
-
-    public function disableDegrees($id)
-    {
-        $degree = Degree::find($id);
-        $degree->disable();
-        return redirect('/degrees');
-    }
-
-    public function activeDegrees($id)
-    {
-        $degree = Degree::find($id);
-        $degree->enable();
-        return redirect('/degrees');
-    }
-
-    public function trashDegrees()
-    {
-        $search = request()->query('search');
-
-        if ($search) {
-            $degree = Degree::where('name', 'LIKE', "%{$search}%")
-                ->where('state', 0)
-                ->paginate(10);
-        } else {
-            $degree = Degree::where('state', 0)->paginate(10);
-        }
-
-        return view('resourcesJV.degrees.trashDegrees', ['degree' => $degree]);
-    }
-
-    public function editDegrees(DegreeRequest $request, $id)
-    {
-        $degree = Degree::find($id);
-
-        if (!$degree) {
-            return redirect('/degrees')->with('error', 'Grado no encontrado.');
-        }
-
-        $degree->update($request->validated());
-
-        return redirect('/degrees')->with('success', 'Grado actualizado correctamente.');
-    }
-
-
-
-
-    //FUNCTIONS SECTIONS
-
-    public function listSections()
-    {
-        $search = request()->query('search');
-
-        if ($search) {
-            $sections = Sections::where('name', 'LIKE', "%{$search}%")
-                ->where('state', 1)
-                ->paginate(10);
-        } else {
-            $sections = Sections::where('state', 1)->paginate(10);
-        }
-
-        return view('resourcesJV.sections.listSections', ['sections' => $sections]);
-    }
-
-    public function createSections(DegreeRequest $request)
-    {
-        $sections = Sections::create($request->validated());
-        return redirect('/sections');
-    }
-
-    public function disableSections($id)
-    {
-        $sections = Sections::find($id);
-        $sections->disable();
-        return redirect('/sections');
-    }
-
-    public function activeSections($id)
-    {
-        $sections = Sections::find($id);
-        $sections->enable();
-        return redirect('/sections');
-    }
-
-    public function trashSections()
-    {
-        $search = request()->query('search');
-
-        if ($search) {
-            $sections = Sections::where('name', 'LIKE', "%{$search}%")
-                ->where('state', 0)
-                ->paginate(10);
-        } else {
-            $sections = Sections::where('state', 0)->paginate(10);
-        }
-
-        return view('resourcesJV.sections.trashSections', ['sections' => $sections]);
-    }
-
-    public function editSection(SectionsRequest $request, $id)
-    {
-        $section = Sections::find($id);
-
-        if (!$section) {
-            return redirect('/sections')->with('error', 'Grado no encontrado.');
-        }
-
-        $section->update($request->validated());
-
-        return redirect('/sections')->with('success', 'Grado actualizado correctamente.');
-    }
-
-
-
-
-
-    //FUNCTIONS COURSES
-
     public function listCourses()
     {
         try {
