+++ resources/views/resourcesJV/degrees/createDegrees.blade.php
@@ -1,5 +1,5 @@
 <div class="p-4 md:p-5">
-    <form class="space-y-4" action="/newDegree" method="POST">
+    <form class="space-y-4" action="degrees/new" method="POST">
         @csrf
         <div>
             <label for="email" class="block mb-2 text-sm font-medium text-gray-900">Nombre del Grado</label>
